Actual proper initialization in Pen, and more errors from main ()
A simply constructed Pen doesn't have an actual known Color or Type,
so both of those are Error and Unknown, respectively.

Also, log a DEBUG message if the job file can't actually be opened
or read from. I don't think the correct action is to fall back to
stdin, so we return 1 after.