Remove unused ColorInfo struct from Pen
This was used for a lookup array. The array was removed because I
feared I would forget everything had to be in the order the Color
enum was in, so I now do everything in the Pen constructor. As a
result, the ColorInfo struct was never used, so it's gone now.