Pen: Throw exceptions with no need for public Color or Type
Color and Type were used in exists () to discover whether or not a
Pen was valid. We now check this in the constructor and throw an
exception should it be invalid.

This also gets rid of a disgusting new Pen (...) followed by a call
to exists () on it and then deleting it if it was false.