Better identification of invalid Pens
Someday, I may want to use this for more HP printer models, where
Pens aren't necessarily invalid if they have no color. Currently,
any marking Pen without a color value is invalid.

This works fine for the DeskJet it was developed for, but will need
additional cases in the future for other models. Besides that, it
should work.