Iterator support is here, but I deleted other useful things
This no longer compiles, as I've accidentally deleted the
copy constructor and destructor implementations. Whoops.

The iterator works, though, but it's repulsive. begin () and
end () return the same DeskJet3600 object, and operator== just
compares curPen to numPens. This isn't great. How would we do
this another way? I thought of a Pens class that implements
Pen storage and is itself an iterator, but now we expose that
everywhere. That said, it encapsulates holding Pens, and we
get it from a DeskJet3600, so maybe that's okay?