range-for used more often by DeskJet3600 users
The old way is gone now.

I'm still debating making the Pens object an actual way to store a
group of Pens. This makes it potentially expensive to create the
begin and end iterators (what if everything gets copied?)
Presently, Pens's constructors take the array as a pointer along
with the array's length and the member we want to begin at for
iteration.