README.md: No longer supporting a forced hpmud build
We now encourage the reader to use the hp-levels utility from
their distribution's HPLIP build. Should there be outcry, I might
add a script for use during build setup that uses HPLIP's config
file to locate the hpmud library. For now, it's probably best not
to depend on a library that normally doesn't get linked to
anything outside HP's tree.