Backend: Correct exception handler
C++98 doesn't seem to have the && operator for extending the
lifetime of the referenced string (we're just returning them from
a method,) so we copy it in the exception handler. This should
only incur cost in the case of an error, and the cost is likely
less than actually attempting to read the status.