amend! Backend now has send( ) and receive( ) methods
We can now do Backend-specific things when we send and receive
from the printer. This way, for example, we can make sure output
is flushed to the printer.

The receive( ) methods are pure virtual, as that involves calling
very backend-specific functions.

send( ) has a default method that writes its buffer to standard
output, then flushes it. This is the official CUPS way to do it
(or will be until CUPS 3.)