Minor formatting; silence warnings in Backends
When creating objects or using built-in C++ features, we insert a
space between the feature or object name (like "if" or "Backend")
and omit said space in function/method calls. This might be
stupid, but it's what I've settled on, and now it's a little more
consistent in DeskJet3600.cpp.

We will eventually support receiving from a Backend, but for now,
only the functions exist. They naturally take parameters we aren't
doing anything with for now, so we cast them void.